backend/requirements.txt
fastapi
uvicorn
praw
snscrape
newspaper3k
feedparser
textblob
redis
python-dotenv
